requirements.txt
flask
gunicorn
openai>=1.0.0
requests
line-bot-sdk
deep-translator
langdetect
